fix build error
DWORD: unsigned long -> std::uint32_t, for 64 bit & non-Windows compatibility. unlike Linux, long is still 32 bit in 64bit Windows

+++ Extensions/_DeLib/NeoStr.h
@@ -2775,7 +2775,7 @@ public:
 									if ((controlParam[0] == L'#')
 										|| (controlParam[0] == L'0'
 										&& (controlParam[1] == L'x' || controlParam[1] == L'X'))) {
-										return _h2d(controlParam.data(), controlParam.size());
+										return (DWORD)_h2d(controlParam.data(), controlParam.size());
 									}
 
 									// dec
